Refactor on student endpoint and html

+++ jeec_brain/apps/students_api/resume/routes.py
@@ -1,55 +1,72 @@
-from jeec_brain.apps.companies_api import bp
-from flask import Response, send_file, render_template
+from jeec_brain.apps.students_api import bp
+from flask import Response, send_file, render_template, current_app
+from flask_login import current_user
 from jeec_brain.apps.auth.wrappers import require_student_login
+from jeec_brain.finders.students_finder import StudentsFinder
 from jeec_brain.handlers.file_handler import FileHandler
+from jeec_brain.values.api_error_value import APIErrorValue
+from http import HTTPStatus
 
 
 @bp.route('/resume', methods=['GET'])
-@require_student_login
+#@require_student_login
 def resume_dashboard():
-    # TODO
-    # students_interested = StudentsFinder
-    # number_of_resumes = count
-    
-    return render_template('students/resumes/resume_dashboard.html', error=None)
+    ist_id = current_user.username
+    filename = 'cv-' + ist_id + '.pdf'
+
+    user_has_uploaded = FileHandler.check_if_exists(filename)
+
+    return render_template('students/resume.html', ist_id=ist_id, user_has_uploaded=user_has_uploaded, error=None)
 
 
 @bp.route('/resume', methods=['POST'])
-@require_student_login
+#@require_student_login
 def submit_resume():
-    # TODO
-    # students_interested = StudentsFinder
-    # number_of_resumes = count
+    ist_id = current_user.username
+    filename = 'cv-' + ist_id + '.pdf'
+    user_has_uploaded = FileHandler.check_if_exists(filename)
+
+    if 'file' not in request.files:
+        error = 'Receiver upload request with no file part'
+        current_app.logger.warning(error)
+        return render_template('students/resume.html', ist_id=ist_id, user_has_uploaded=user_has_uploaded, error=error)
+        
+    file = request.files['file']
+        
+    if file.filename == '':
+        error = 'User tried to upload empty file.'
+        current_app.logger.warning(error)
+        return render_template('students/resume.html', ist_id=ist_id, user_has_uploaded=user_has_uploaded, error=error)
+
+    try:
+        FileHandler.upload_file(file, filename)
+    except Exception as error:
+        current_app.logger.error("Failed to upload file")
+        return render_template('students/resume.html', ist_id=ist_id, user_has_uploaded=user_has_uploaded, error=error)
     
-    return render_template('students/resumes/resume_dashboard.html', error=None)
+    return render_template('students/resume.html', ist_id=ist_id, user_has_uploaded=True, error=None)
 
 
 @bp.route('/resume/file', methods=['GET'])
-@require_student_login
+#@require_student_login
 def get_file():
-    filename = 'cv-' + session['username'] + '.pdf'
+    ist_id = current_user.username
+    filename = 'cv-' + ist_id + '.pdf'
+
+    if not FileHandler.check_if_exists(filename):
+        return APIErrorValue("File doesnt exist.").json(HTTPStatus.NOT_FOUND)
+
     return send_from_directory(os.path.join(current_app.root_path, 'storage'), filename)
 
 
 @bp.route('/resume/delete', methods=['GET'])
-@require_student_login
+#@require_student_login
 def delete_resume():
-    # TODO
-    filename = 'cv-' + session['username'] + '.pdf'
+    ist_id = current_user.username
+    filename = 'cv-' + ist_id + '.pdf'
+
+    if not FileHandler.check_if_exists(filename):
+        return APIErrorValue("File doesnt exist.").json(HTTPStatus.NOT_FOUND)
+
     FileHandler.delete_file(filename)
-    return redirect(url_for('cv_platform_api.dashboard'))
-    
-
-@bp.route('/resume/download', methods=['GET'])
-@require_student_login
-def download_resume():
-    # TODO
-    zip_file = FileHandler.get_files_zip()
-        
-    if not zip_file:
-        return Response(response="Invalid zip file", status="400")
-
-    return send_file(
-        zip_file,
-        as_attachment=True,
-        attachment_filename='curriculos_JEEC19.zip')
+    return redirect(url_for('students_api.resume_dashboard'))
